+++ GameCPP/GameCPP/TranspositionTable.cpp
@@ -4,14 +4,14 @@
 #include "TranspositionTable.h"
 
 void
-TranspositionTable::insert(long zkey, BreakthroughMove move, int depth, int score, int min, int max) {
+TranspositionTable::insert(long long zkey, BreakthroughMove move, int depth, int score, int min, int max) {
 	TableEntry entry(zkey, move, depth, score, min, max);
 	int index = (int)(zkey%TABLE_SIZE);
 	table[index] = entry;
 }
 
 TableEntry&
-TranspositionTable::lookup(long zkey) {
+TranspositionTable::lookup(long long zkey) {
 	int index = (int)(zkey%TABLE_SIZE);
 	if (table[index].getKey() == zkey) {
 		return table[index];
